Refactor ActualBudget node to resource+operation pattern
Split the flat operation dropdown into a Resource (Transaction/Budget)
selector with per-resource operation dropdowns and dedicated action
modules, mirroring n8n's standard node structure. Pure refactor: the
4 existing operations behave identically, just organized to scale as
more resources are added.
